3-coffee-shop-full-stack-iam/backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
astroid==2.2.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11; python_version >= '3'
click==7.0
cryptography==36.0.1; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
ecdsa==0.13.2
flask-cors==3.0.8
flask-sqlalchemy==2.4.0
flask==1.0.2
future==0.17.1
idna==3.3; python_version >= '3'
isort==4.3.18
itsdangerous==1.1.0
jinja2==2.10.1
lazy-object-proxy==1.4.0
markupsafe==1.1.1
mccabe==0.6.1
oauthlib==3.2.0; python_version >= '3.6'
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.3.1
pyjwt==2.3.0; python_version >= '3.6'
pylint==2.3.1
python-jose-cryptodome==1.3.2
python-jose==3.3.0
python3-openid==3.2.0
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.8; python_version >= '3.6' and python_version < '4'
six==1.12.0
social-auth-core[openidconnect]==4.2.0
sqlalchemy==1.3.3
typed-ast==1.4.1
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1
wrapt==1.11.1
